refactor(ui): move header components out of Browse2
