+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/common/tabularListViews/CollapseHeader.tsx
@@ -4,8 +4,8 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import {Button} from '../../../Button';
-import {Tooltip} from '../../../Tooltip';
+import {Button} from '../../../../../../Button';
+import {Tooltip} from '../../../../../../Tooltip';
 
 type CollapseHeaderProps = {
   headerName: string;
